requirements.txt
fastapi
uvicorn
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
diffusers==0.21.4
transformers==4.36.2
huggingface_hub==0.19.4
numpy==1.26.4
Pillow
segment-anything
